+++ R/plot_meta_production.R
@@ -5,14 +5,10 @@
 #'
 #' @param result \code{meta_rn} object simulated with \code{run_meta}.
 #' @param lag Logical if TRUE, the difference to the previous timestep is returned.
-#' @param turnover Logical if TRUE, the turnover is returned.
 #' @param base_size Numeric to specify base font size.
 #'
 #' @details
 #' Plot the production per meta ecosystem.
-#'
-#' If \code{turnover = TRUE}, the turnover is calculated defined as
-#' turnover = biomass / production.
 #'
 #' @examples
 #' \dontrun{
@@ -23,16 +19,23 @@
 #' @rdname plot_meta_production
 #'
 #' @export
-plot_meta_production <- function(result, lag = TRUE, turnover = FALSE, base_size = 10) {
+plot_meta_production <- function(result, lag = TRUE, base_size = 10) {
 
   # calculate production
-  production <- get_meta_production(result = result, lag = lag, turnover = turnover)
+  production <- summarize_meta(result = result, what = "production",
+                               lag = lag)[["production"]]
 
   # remove NA rows (first row)
   production <- production[stats::complete.cases(production), ]
 
   # better id col nameing
   production$meta <- paste0("Meta_", production$meta)
+
+  production <- stats::reshape(production, direction = "long",
+                               v.names = "value", varying = c("bg_production", "ag_production", "ttl_production"),
+                               timevar = "part", times = c("bg_production", "ag_production", "ttl_production"),
+                               new.row.names = 1:(nrow(production) * 3),
+                               idvar = "timestep", ids = unique(production$timestep))
 
   # create title
   plot_title <- paste0("Total time        : ", result$max_i, " iterations [",
@@ -43,7 +46,7 @@ plot_meta_production <- function(result, lag = TRUE, turnover = FALSE, base_size
   # create plot
   gg_prod <- ggplot2::ggplot(data = production) +
     ggplot2::geom_line(ggplot2::aes(x = timestep, y = value, col = factor(meta))) +
-    ggplot2::facet_wrap(. ~ part, scales = "free_y", nrow = 2, ncol = 1) +
+    ggplot2::facet_wrap(. ~ part, scales = "free_y", nrow = 3, ncol = 1) +
     ggplot2::scale_color_viridis_d(name = "", option = "C") +
     ggplot2::labs(x = "Timestep", y = "Biomass production", title = plot_title) +
     ggplot2::theme_classic(base_size = base_size) +