+++ data-raw/result_readme.R
@@ -15,7 +15,7 @@ n <- 3
 starting_values$pop_n <- rep(x = 5, n = n)
 
 # create 5 reef cells in center of seafloor
-reefs <- matrix(data = c(-1, 0, 0, 1, 1, 0, 0, -1, 0, 0), ncol = 2, byrow = TRUE)
+reef <- matrix(data = c(-1, 0, 0, 1, 1, 0, 0, -1, 0, 0), ncol = 2, byrow = TRUE)
 
 # setup extent and grain
 dimensions <- c(50, 50)
@@ -36,7 +36,8 @@ parameters$nutrients_output <- 0.1
 
 parameters$move_residence <- max_i / 10
 
-stable_vals <- arrR::get_stable_values(starting_values = starting_values,
+stable_vals <- arrR::get_stable_values(bg_biomass = starting_values$bg_biomass,
+                                       ag_biomass = starting_values$ag_biomass,
                                        parameters = parameters)
 
 starting_values$nutrients_pool <- stable_vals$nutrients_pool
@@ -47,7 +48,7 @@ nutr_input <- sim_nutr_input(n = n, max_i = max_i, input_mn = stable_vals$nutr_i
                              freq_mn = 5, variability = 0.5)
 
 metasyst <- setup_meta(n = n, max_i = max_i, dimensions = dimensions, grain = grain,
-                       reefs = reefs, starting_values = starting_values, parameters = parameters)
+                       reef = reef, starting_values = starting_values, parameters = parameters)
 
 result_readme <- run_meta(metasyst = metasyst, nutr_input = nutr_input,
                           parameters = parameters, movement = "attr",