+++ tests/testthat/test-setup_meta.R
@@ -1,6 +1,6 @@
 # create vector with names
 present_names <- c("seafloor", "fishpop", "n", "fishpop_attributes",
-                   "starting_values", "parameters", "reefs", "extent",
+                   "starting_values", "parameters", "reef", "extent",
                    "grain", "dimensions")
 
 test_that("setup_meta returns meta_syst", {
@@ -36,8 +36,8 @@ test_that("setup_meta returns correct information", {
 
   expect_equal(object = metasyst$n, expected = n)
 
-  expect_equal(object = metasyst$extent,
-               expected = raster::extent(dimensions / 2 * c(-1, 1, -1, 1)))
+  expect_equal(object = terra::ext(metasyst$extent),
+               expected = terra::ext(dimensions / 2 * c(-1, 1, -1, 1)))
 
   expect_equal(object = metasyst$grain, expected = c(grain, grain))
 
@@ -46,8 +46,8 @@ test_that("setup_meta returns correct information", {
 
   expect_equal(object = metasyst_nr$n, expected = n)
 
-  expect_equal(object = metasyst_nr$extent,
-               expected = raster::extent(dimensions / 2 * c(-1, 1, -1, 1)))
+  expect_equal(object = terra::ext(metasyst_nr$extent),
+               expected = terra::ext(dimensions / 2 * c(-1, 1, -1, 1)))
 
   expect_equal(object = metasyst_nr$grain, expected = c(grain, grain))