+++ NEWS.md
@@ -3,12 +3,12 @@
 * Add `calc_variability()` and `sample_variability()` function
 * Add `sim_nutr_input()` function
 * Add `get_input_df()` function
-* Add `filter_nutr_input()` function
+* Add `filter_meta()` function
 * Simplified code for `plot.meta_rn()`
 * Add `get_meta_production()` and `plot_meta_production()`
-* Add `filter_metarn()`
 * Remove `timestep` argument from many functions
-* Combined `filter_nutr_input` and `filter_metarn` to `filter_meta`
+* Print time-stamp during `run_meta()`
+* Add `summarize_meta()` function and delete `get_meta_production`
 
 # meta.arrR 0.2.1
 * Adding `testthat` functions for all R functions